change type

+++ src/lib.rs
@@ -10,8 +10,8 @@ pub fn three() -> u32 {
     3
 }
 
-pub fn four() -> u32 {
-    4
+pub fn four() -> f64 {
+    4.0
 }
 
 #[cfg(test)]
